+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.5.0] - unreleasd
+### Added
+- Sequenceimage: Add download function
+
 ### Changed
 - CaptureButton created and integrated into CameraView 
 - CameraView is no longer locked when a sequence is running. Capturing only is not possible